firs draft of GFF3 format for small RNA outpout
Hi all, quick draft for GFF3 format. Please, contribute at your pace, I will try to add something every day.

Keep in mind this is for the output of a pipeline, so we know there will be bias to methodology, but the idea is to put enough information to be able to re-analyze or filter sequences using features described here. As well, it would be a proxy for downstream analysis or packages.

cc @lpantano @gurgese @ThomasDesvignes @mhalushka @mlhack @keilbeck @BastianFromm @ivlachos @TJU-CMC